Timer now correct, started favourites and replay

+++ app/src/main/java/pt/isel/pdm/gomoku_ee/Game/GameView.kt
@@ -44,7 +44,7 @@ import java.util.UUID
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun GameView(game: Game, timer: Int,onMainRequested: () -> Unit = {}, onUpdate: (GamePlayInputModel) -> Unit = {}) {
+fun GameView(game: Game, timer: Long, onMainRequested: () -> Unit = {}, onUpdate: (GamePlayInputModel) -> Unit = {}) {
     var fav by remember { mutableStateOf(false) }
     Column(
         modifier = Modifier
